Introduce dynamic scoping.
Scope is no longer an interface, but a class which objects represent the current state of the program.

Callables don't have predefined scopes, rather the scope is passed to them upon execute() and debug() methods.

ScopeInstructions create an 'innerScope' as a child of the scope which was passed to the executable method and pass it to their instructions.

Block now has two execute() and debug() methods, one of each for returning the created innerScope when executed from Macchiato.